refactor(#295): estrae ContrastBadge condiviso e corregge chiavi i18n
ContrastBadge duplicato in StyleGuide e SettingsModal centralizzato in components/ui.
Corrette le chiavi colorScheme_dark/light mancanti del prefisso settings.
